Remodeling - Work on service/app model (#63)
* fix long tests in sql state bootstrapping.

Logic around rendezvous after transactions have been completed was
causing the tests to wait the full timeout :(

* Use a separate limit for each client connection, rather than sharing.

moar rational thread pooling for local servers. moar rational params for
default limits.

* test fixins n' stablizers

* Clean up tech debt on Context impl.

3 failing tests only.  interim commit

* refactor to manage context membership

back off some changes

* fix assembly test

* fix sql state bootstrapping testing

genesis wasn't filtering out the test subject in the transactioneers and
so was flapping pretty wildly, as the test subject was banging until
bootstrapped, and then the only transactioneer - lol.  So would take
forever.

* refactor liquibase schemas to a common module.

This is a work around until JOOQ is capable of working with current H2
version.

Clean up Stereotomy Resolver.

Clean up liquibase change log organization for modular reuse.  probably
will need to go into full blown package space naming.

Clean up model schema, link with stereotomy

* Favor completion over throughput for lifecycle bootstrapping tests

* mea culpa

* update

* make KEL.append() async.

Clean up some cruft in stereotomy

* use supplier for genesis data

* flesh out grpc Resolver/Binder. Rename stereotomy_kerl schema

* misc

* Move to overlap of view reconfiguration.

Fix age batching q limit

* stabilize

* better error msg

* fix batch limit in producer.

Apparently I have it correct now.  No dropped batches when starting
assembly.

* update grpc version

* Refactor Genesis Assembly

Genesis block now assembled in a rational manner, no queries for vm, as
there is no ability to validate them and we know the initial set.

Add Foundation and FoundationSeal for trust establishment of genesis.
Joins now contain the KEL of the founding members of the genesis
committee.

* generic prefix for genesis reContext

* Tighten up Stereotomy API

* moar cleanup, honing in on delegated rotation/inception in direct mode

* Clean up AttachmentEvent & Co.

Overhaul stereotomy schema for cascade delete constraints, etc.

* remove unused columns

* still continuing

* Revamp delegated establishment, phase 1

Delegation linkage to delegator event is now handled by attachements

Phase 1 is the revamp of the delegated establishment events and
validation

* Direct mode delegation working for MVLog KERL version

* Append attachment for UniKERL

* don't parallelize sql-state testing

* Bulk append. Refactor out delegated direct mode logic in incept and rot

* commentary

* Refactor to use ULong instead of Long for block height

Needed to be done sooner, rather than later.

* uint64

lol.  like it matters ;)  Still...

* Use ULong for event sequence number

Sooner > later ;)

* fix type issue. accepted by ide, rejected by command line compiler. lol

* minor waxing.

Seems clear my failures for these two are caused by random interactions
of eclipse with these during a command line maven build ;)

Close all the editors, and don't touch the IDE during the CLI build.

* Update to H2 2.206

Mods ported

* first smoke test of mutator testing

* Update to JOOQ 3.16.0

Use only one version of H2 for gen and runtime.

* back off JOOQ 3.16.0

lol

* add batched transaction testing

* Fix Sharded Kerl, using batch

* update schemas to use identity, as is the fashion

* Simple Emulation for a SQL State Machine Shard.

Sutable for development, testing, etc.  Should have done this years ago
- lol

* lotta cleanup.

normalize default H2 UPPER_CASE default behavior. Eliminate YML as a
liquibase format.  Geebus.

Add changelog resources from defined {path -> URL} mappings.

Fix a lot of stupid stuff in Mutator around Call, BatchedTransactions
(yay!).

Model now depends on everything (lol). Intialize model genesis
transaction (i.e. the combined schemas liquibase migrations), test
ShardedKERL for basic functionality.

* Async Delphinius API

First cut at at CHOAM mutation for Delphinius Oracle in the
ShardedOracle implementation

(tl;dr: that test currently fails)

* import cleanup

* Update README.md

* fix sharded oracle test

* sync

* Remove MVLog based KERL, just use in memory maps.

Serialization never worked anyway.

* Stereotomy cleanup. Get attachments for event, get KERL of identifier.

* Optional api conformance

* forgot these

* Spring cleaning

* Access KERL on controlled identifier

* Paramaterize the MvStore for CHOAM

* Work on defining the bootstrapping of the Node

more work on Stereotomy, fleshing out the rest.

* Flesh out some remaining impls, etc.

* Add type variables for identifier construction, etc.

Lotta hair on that ;)

Integrate fake provisioning of the member.

* typing

* moar cleanup.

Generisize identifier creation in Sterotomy - prolly needs a bit more
work, though.

Remove unneeded publicKey() from Member - lot of fallout from that

* Node simple smoke test.

Bootstrapping with KERL.  Some fixins

Test simply boots and waits.

* remove unneeded interface

* Update to H2 2.1.210

* Update to JOOQ 3.16.2

Remove need to use a seperate version of H2 for generation

* update bouncy castle 1.70, Dropwizard 2.0.28, Jackson 2.13.1

* First cut at non multi-tenant demo of RBAC service

* flesh out implementations

* combine. implement check

* use DELETE for remove/delete

* Refactor to seperate runtime parameters from constructable params

* Moar split of runtime/configtime. Can now use for YAML dropwizard config

* Further fleshing out the demo API

lol.  Proto over HTTP for the KERI resolver

+++ choam/src/main/java/com/salesforce/apollo/choam/support/CheckpointAssembler.java
@@ -20,6 +20,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.IntStream;
 
 import org.h2.mvstore.MVMap;
+import org.joou.ULong;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -53,11 +54,11 @@ public class CheckpointAssembler {
     private final DigestAlgorithm                           digestAlgorithm;
     private final double                                    fpr;
     private final List<Digest>                              hashes    = new ArrayList<>();
-    private final long                                      height;
+    private final ULong                                     height;
     private final SigningMember                             member;
     private final MVMap<Integer, byte[]>                    state;
 
-    public CheckpointAssembler(long height, Checkpoint checkpoint, SigningMember member, Store store,
+    public CheckpointAssembler(ULong height, Checkpoint checkpoint, SigningMember member, Store store,
                                CommonCommunications<Terminal, Concierge> comms, Context<Member> context,
                                double falsePositiveRate, DigestAlgorithm digestAlgorithm) {
         this.height = height;
@@ -90,7 +91,7 @@ public class CheckpointAssembler {
         });
         CheckpointReplication.Builder request = CheckpointReplication.newBuilder()
                                                                      .setContext(context.getId().toDigeste())
-                                                                     .setCheckpoint(height);
+                                                                     .setCheckpoint(height.longValue());
         request.setCheckpointSegments(segmentsBff.toBff());
         return request.build();
     }
