+++ stereotomy/src/main/java/com/salesforce/apollo/stereotomy/event/protobuf/KeyEventImpl.java
@@ -10,6 +10,8 @@ import static com.salesforce.apollo.crypto.QualifiedBase64.digest;
 import static com.salesforce.apollo.crypto.QualifiedBase64.signature;
 import static com.salesforce.apollo.stereotomy.identifier.QualifiedBase64Identifier.identifier;
 
+import org.joou.ULong;
+
 import com.google.protobuf.ByteString;
 import com.salesfoce.apollo.stereotomy.event.proto.EventCommon;
 import com.salesfoce.apollo.stereotomy.event.proto.Header;
@@ -74,8 +76,8 @@ abstract public class KeyEventImpl implements KeyEvent {
     }
 
     @Override
-    public long getSequenceNumber() {
-        return header.getSequenceNumber();
+    public ULong getSequenceNumber() {
+        return ULong.valueOf(header.getSequenceNumber());
     }
 
     @Override
